% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseGen.R
\name{get.dgp}
\alias{get.dgp}
\title{get.dgp}
\usage{
get.dgp(n, d, pos = 0.01, minBV = 0.03)
}
\arguments{
\item{n, }{sample size}

\item{d, }{set to 4 only}

\item{pos, }{a small value to make sure prop scores are between in (pos, 1 - pos)}

\item{minBV, }{minimum blip variance}
}
\value{
a sample DF, the true average treatment effect, ATE0 and blip variance
BV0, the sample pscores, PGn, the sample true blips, blip_n, the sample 
true prob of death under treatment, PQ1n, and prob of death under control
PQ0n
}
\description{
randomly creates a dgp, for now only 4 dimensional of confounders, binary 
treatment.
}
\examples{
# getting a random sample from a randomly drawn dgp 
dgp = get.dgp(n = 1000, d = 4)

# population blip variance
dgp$BV0
# population average treatment effect
dgp$ATE0
# sample proportion with treatment
mean(dgp$DF$A)
# sample proportion who died
mean(dgp$DF$Y)
# min sample pscore
min(dgp$PGn)
# max sample pscore
max(dgp$PGn)
# sample dataframe
head(dgp$DF)
# histogram of blips
hist(dgp$blip_n)

}
