% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.inference.R
\name{IC.beta}
\alias{IC.beta}
\title{IC.beta}
\usage{
IC.beta(X, Y, Qk)
}
\arguments{
\item{X, }{design matrix}

\item{Y, }{a binary vector of outcomes}

\item{Qk, }{predictions}
}
\value{
a matrix that is the IC for the beta coefficients of the logistic model
}
\description{
mainly a helper function to get IC for the betas of logistic regression
}
