% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{LR.inference}
\alias{LR.inference}
\title{LR.inference}
\usage{
LR.inference(W, A, Y, Qform, alpha = 0.05, simultaneous.inference = FALSE)
}
\arguments{
\item{W, }{matrix or data.frame of covariates}

\item{A, }{a binary vector of treatment assignments}

\item{Y, }{a binary vector of outcomes}

\item{Qform, }{a formula for Y in terms of the covariates as input in glm}

\item{alpha, }{significance level for the CI. 0.05 is default}

\item{simultaneous.inference, }{TRUE if user wants simultaneous confidence
bounds for both ATE and blip variance.  default is FALSE}
}
\value{
if simultaneous.inference is specified as TRUE then will return a vector giving
pt estimate, left and right bound for blip variance, blip variance simultaneous CI, ATE, 
and ATE simultaneous CI.  Otherwise gives pt estimate, left and right bound for blip variance
and ATE.
}
\description{
Function that gives inference for logistic regression plug-in
estimators of ATE and Blip Variance.
}
\examples{
# using built-in package functions, g0_linear and Q0_1 to specify
# pscore and outcome model probabilities
data = gendata(1000, g0_linear, Q0_1)

Qform = formula("Y ~ A*W1 + W2 + W3 + A*W4 + W4 + A")

# specifying the covariates, treatment and outcome
W = data[,2:5]
A = data$A
Y = data$Y

info = LR.inference(W=W,A=A,Y=Y,Qform=Qform)
info
}
