% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.inference.R
\name{LR.TSM}
\alias{LR.TSM}
\title{LR.TSM}
\usage{
LR.TSM(W, A, Y, Qform, setA, alpha = 0.05)
}
\arguments{
\item{W, }{matrix or data.frame of covariates}

\item{A, }{a binary vector of treatment assignments}

\item{Y, }{a binary vector of outcomes}

\item{Qform, }{a formula for Y in terms of the covariates as input in glm}

\item{setA, }{the number you are setting treatment to}

\item{alpha, }{significance level for the (1-alpha)100 percent CI's. 0.05 is default}
}
\value{
a list with components CI and IC, giving the influence curve approximation
for the logistic regression plug-in estimator for the TSM with A set to setA
}
\description{
Function that gives inference for logistic regression plug-in
estimators of ATE and Blip Variance.
}
\examples{

# using built-in package functions, g0_linear and define Q0_linear to specify
# pscore and outcome model probabilities
g0_linear
Q0_linear = function(A,W1,W2,W3,W4) plogis(A + W1 + W2 + A*(W3 + W4))

# get a randomly drawn dataframe under the specified model
data = gendata(1000, g0_linear, Q0_linear)
big = gendata(1e6, g0_linear, Q0_linear)
# get the truth setting A to 1
setA = 1
truth = mean(with(big, Q0_linear(setA,W1,W2,W3,W4)))

# well-specified model
Qform = formula("Y ~ W1 + W2 + A*(W3 + W4)")

# specifying the covariates, treatment and outcome
W = data[,2:5]
A = data$A
Y = data$Y

# should cover each truth 95 percent of the time.
info = LR.TSM(W=W,A=A,Y=Y,Qform=Qform, setA = 1, alpha = .05)
# get CI
info$CI
# get influence curve
info$IC


}
