% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.inference.R
\name{IC.beta}
\alias{IC.beta}
\title{IC.beta}
\usage{
IC.beta(data, OC = NULL, Ynode, Qform, verbose = FALSE,
  parallelize = FALSE)
}
\arguments{
\item{Qform, }{a formula for Y in terms of the covariates as input in glm}

\item{W, }{matrix or data.frame of covariates}

\item{A, }{a binary vector of treatment assignments}

\item{Y, }{a binary vector of outcomes}
}
\value{
a list with elements IC_beta and fit, the glm fit object.
}
\description{
computes IC for beta coefficients of logistic regression
mainly a helper function.
}
