% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.inference.R
\name{long.TSM}
\alias{long.TSM}
\title{long.TSM}
\usage{
long.TSM(data, Ynodes, Anodes, formulas, setA)
}
\arguments{
\item{data, }{data.frame of variables in time ordering from left to right}

\item{Ynodes, }{character vector of time-ordered Ynodes}

\item{Anodes, }{character vector of time-ordered Ynodes}

\item{formulas, }{list of formulas for the conditional means}

\item{setA}{the value to which you intervene on A,vector of length that of Anodes}

\item{alpha}{significance level for two-sided CI.}
}
\value{
a list with elements, CI for the confidence interval and  IC for the 
influence curve.
}
\description{
computes IC and psi for logistic regression plug-in estimator of treatment
specific mean for survival data in wide form. IN DEVELOPMENT
}
