% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.inference.R
\name{LR.TSM}
\alias{LR.TSM}
\title{LR.TSM}
\usage{
LR.TSM(data, Ynode, Anode, Qform, setA, alpha = 0.05)
}
\arguments{
\item{Qform, }{a formula for Y in terms of the covariates as input in glm}

\item{setA}{the value to which you intervene on A, the treatment}

\item{alpha}{significance level for two-sided CI}

\item{W, }{matrix or data.frame of covariates}

\item{A, }{a binary vector of treatment assignments}

\item{Y, }{a binary vector of outcomes}
}
\value{
a list with elements IC for the influence curve and CI for 
the confidence interval
}
\description{
computes IC for logistic regression plug-in estimator of treatment
specific mean and the corresponding wald confidence interval
}
