% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.inference.R
\name{LR.TSM}
\alias{LR.TSM}
\title{LR.TSM}
\usage{
LR.TSM(W, A, Y, Qform, setA, alpha = 0.05)
}
\arguments{
\item{W, }{matrix or data.frame of covariates}

\item{A, }{a binary vector of treatment assignments}

\item{Y, }{a binary vector of outcomes}

\item{Qform, }{a formula for Y in terms of the covariates as input in glm}

\item{setA, }{the number you are setting treatment to}

\item{alpha, }{significance level for the (1-alpha)100 percent CI's. 0.05 is default}
}
\value{
a list with components CI and IC, giving the influence curve approximation
for the logistic regression plug-in estimator for the TSM with A set to setA
}
\description{
Function that gives inference for logistic regression plug-in
estimators of ATE and Blip Variance.
}
