% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{sim_cv}
\alias{sim_cv}
\title{sim_cv}
\usage{
sim_cv(n, g0, Q0, SL.library, SL.libraryG, method = "method.NNLS",
  cv = TRUE, V = 10, SL = 10L, gform, Qform, estimator, dgp = NULL,
  gn = NULL)
}
\arguments{
\item{n, }{sample size}

\item{g0, }{treatment mechanism function, call g0_linear to see the format}

\item{Q0, }{outcome model function, call Q0 linear to see the format--WILL
be more generalized but for now rather limited}

\item{SL.library, }{SuperLearner library for outcome predictions}

\item{SL.libraryG, }{SuperLearner Library for treatment mechanism}

\item{method, }{SuperLearner meta fitting method}

\item{cv, }{set to TRUE for CV-TMLE}

\item{V, }{number of folds for the CV tmle}

\item{SL, }{number of folds for each superlearner}

\item{gform, }{a linear form to specify for estimating pscore}

\item{Qform, }{a linear form to specify for estimating outcome prediction}

\item{estimator, }{a character vector containing any set of "single 1step" for
one step tmle single param estimates for ATE and blip variance, "single iterative"
for the same with iterative tmle, or "simul 1 step", "simul line", "simul full"
to compute simultaneous estimates and CI's for ATE and blip variance.  line, full
and 1 step are just different targeting methods for tmle}

\item{dgp, }{a list containing an element named DF for the data.frame with A, Y and 
covariates which are named whatever, BV0 and ATE0 for true blip variance and 
average treatment effect respectively.}
}
\value{
a vector with the following elements in this order:
TMLE pt estimates and confidence intervals each with estimate, 
left and right bounds and initial estimates for BV and ATE
Superlearner coefficients and risks for both pscore and outcome
estimation, tmle pt estimates, CI's for BV and ATE for Qform model 
assumed plus initial estimates for these as well as pt estimates and
CI's for BV and ATE using the delta method, ie sandwich estimator
under non-parametric model
}
\description{
Function that simulates data and performs TMLE estimates
data is 4 covariates and binary treatment and outcome.  The covariates
are generated according to the gendata function.
}
\examples{
# basic example with very simple SuperLearner library
SL.library = c("SL.glm", "SL.mean")
SL.libraryG = c("SL.glm", "SL.mean")
n=1000
g0 = g0_linear
Q0 = Q0_trig
gform = formula("A ~.")
Qform = formula("Y ~ A*(W1 + W2 + W3 + W4)")
estimator = c("single iterative")
result = sim_cv(n, g0, Q0, SL.library, SL.libraryG, method = "method.NNLS", 
                  cv = TRUE, V = 2, SL = 2L, gform, Qform, estimator, dgp = NULL)
          

}
